amazon code -> amazon view
we learned that cgi escaping is better than uri escaping, not
going to mess with Addressable since it looks needlessly complex
and CGI probably does what we want.

took a lot of Amazon-specific code out of the app.rb, out and into
Amazon view where it pollutes less.  Next up is BossService view.

created a Document parent view to illustrate relationship between
Amazon and BossService; they are both providers of document, which
should at least provide body; here it's done explicitly in app.rb,
but as long as the document mustache is the one being used and have
"!@@have.nil?" is satisfied, it's presumed that doc got needed info
and calling the "got" method returns it, or potentially calls some
method provider who overrides body in the document view instead of
pulling it directly from @@have.  (This may be a work of fiction.)

{{#got}}{{{body}}}{{/got}} <!-- might be a convenient idiom later.
{{^have}}Missing struct import may signal something wrong{{/have}}